Revert "database quikchanges"
This reverts commit 2e24abf916cd2ed8017ac5b44feb0020c78949c0.

+++ migrations/20240529135958-create-user.js
@@ -38,7 +38,7 @@ module.exports = {
         type: Sequelize.BOOLEAN
       },
       mobileNumber: {
-        type: Sequelize.INTEGER(20)
+        type: Sequelize.INTEGER
       },
       companyName: {
         type: Sequelize.STRING
@@ -47,7 +47,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       whatsapp: {
-        type: Sequelize.INTEGER(20),
+        type: Sequelize.INTEGER,
         defaultValue: null
       },
       facebook: {
